Salesforce - send customer to salesforce contact updates - Move to shopify/customer (renamed from shopify/customers - Update mesa.json and README, rename scripts
