--- shopify/customers/send-to-salesforce-contact/input.js
@@ -1,17 +1,0 @@
-const Mesa = require('vendor/Mesa1.js');
-
-/**
- * A Mesa Script exports a class with a script() method.
- */
-module.exports = new class {
-
-  /**
-   * Mesa Script
-   *
-   * @param {object} payload The payload data
-   * @param {object} context Additional context about this task
-   */
-  script = (payload, context) => {
-    Mesa.output.send(`${context.input.key.replace('in-', 'out-')}`, payload);
-  }
-}
