--- shopify/customers/send-to-salesforce-contact/mesa.json
@@ -1,57 +1,0 @@
-{
-  "name": "Send New Shopify Customers to Salesforce Contact",
-  "version": "1.0.0",
-  "description": "Sends new Shopify customers to Salesforce contacts when they're created.",
-  "documentation": "https://github.com/shoppad/mesa-templates/blob/master/shopify/customers/send-to-salesforce-contact/README.md",
-  "tags": ["customers", "salesforce"],
-  "source": "shopify",
-  "destination": "salesforce",
-  "config": {
-    "inputs": [
-      {
-        "trigger_type": "input",
-        "enabled": true,
-        "type": "shopify_webhook",
-        "name": "Send Shopify Customer to Salesforce Contact",
-        "key": "in-send-shopify-customer-to-salesforce",
-        "trigger_name": "In: Send Shopify Customer To Salesforce",
-        "script": "shopify/customers/send-to-salesforce-contact/input.js",
-        "logging": "error",
-        "topic": "customers/create"
-      }
-    ],
-    "outputs": [
-      {
-        "trigger_type": "output",
-        "enabled": true,
-        "type": "custom",
-        "name": "Send Shopify Customer To Salesforce Contact",
-        "key": "out-send-shopify-customer-to-salesforce",
-        "trigger_name": "Out: Send Shopify Customer To Salesforce",
-        "script": "shopify/customers/send-to-salesforce-contact/output.js"
-      }
-    ],
-    "secrets": [
-      {
-        "name": "salesforce_client_id"
-      },
-      {
-        "name": "salesforce_client_secret"
-      },
-      {
-        "name": "salesforce_username"
-      },
-      {
-        "name": "salesforce_password"
-      },
-      {
-        "name": "salesforce_access_token"
-      }
-    ]
-  },
-  "files": [
-    "shopify/customers/send-to-salesforce-contact/input.js",
-    "shopify/customers/send-to-salesforce-contact/output.js",
-    "shopify/customers/send-to-salesforce-contact/shopify-salesforce-customer-map.js"
-  ]
-}
